Migrade ECDSA sign/verify
